# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustChain - Financial Reputation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  created_at: string;
}

export interface CreateUserBody {
  name: string;
  email: string;
}

/**
 * Map of userId to score (300-850)
 */
export interface AllScores {
  [key: string]: number;
}

export interface ScoreBreakdownItem {
  label: string;
  points: number;
}

export interface VoucherInfo {
  id: string;
  name: string;
  score: number;
  strength: number;
}

export interface UserScoreBreakdown {
  userId: string;
  score: number;
  breakdown: ScoreBreakdownItem[];
  vouchers: VoucherInfo[];
}

export interface Signal {
  id: string;
  user_id: string;
  type: string;
  months_consistent: number;
  created_at: string;
}

export type AddSignalBodyType =
  (typeof AddSignalBodyType)[keyof typeof AddSignalBodyType];

export const AddSignalBodyType = {
  rent: "rent",
  utility: "utility",
  mobile: "mobile",
  default: "default",
} as const;

export interface AddSignalBody {
  userId: string;
  type: AddSignalBodyType;
  months_consistent: number;
}

export interface Vouch {
  id: string;
  voucher_id: string;
  vouchee_id: string;
  strength: number;
  created_at: string;
}

export interface AddVouchBody {
  voucherId: string;
  voucheeId: string;
  strength: number;
}

export interface ExplainBody {
  userId: string;
}

export interface ExplainResponse {
  explanation: string;
}

export interface NetworkNode {
  id: string;
  name: string;
  score: number;
}

export interface NetworkLink {
  source: string;
  target: string;
  strength: number;
}

export interface NetworkData {
  nodes: NetworkNode[];
  links: NetworkLink[];
}

export interface SeedResponse {
  message: string;
  usersCreated: number;
}