# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrustChain - Financial Reputation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scores/{id}:
    get:
      operationId: getUserScore
      tags: [scores]
      summary: Get one user's score and breakdown
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Score with breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserScoreBreakdown"
...
components:
  schemas:
...
    ScoreBreakdownItem:
      type: object
      properties:
        label:
          type: string
        points:
          type: number
      required:
        - label
        - points
    UserScoreBreakdown:
      type: object
      properties:
        userId:
          type: string
        score:
          type: number
        breakdown:
          type: array
          items:
            $ref: "#/components/schemas/ScoreBreakdownItem"
        vouchers:
          type: array
          items:
            $ref: "#/components/schemas/VoucherInfo"
      required:
        - userId
        - score
        - breakdown
        - vouchers
    VoucherInfo:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        score:
          type: number
        strength:
          type: number
      required:
        - id
        - name
        - score
        - strength
...