# lib/api-zod/src/generated/types/voucherInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrustChain - Financial Reputation API
 * OpenAPI spec version: 0.1.0
 */

export interface VoucherInfo {
  id: string;
  name: string;
  score: number;
  strength: number;
}